Move MHLO default ABI generator to native bindings.
The only thing it had special for MHLO is that it flattened tuples
(didn't even model them as actual tuple annotations), but since the
tuples get flattened in the IR anyway, the simpler pass that doesn't
handle tuples is sufficient.

We also put in some defensive assertions around the ABI annotations --
emitting errors if some functions are annotated but others are not. This
might be too strict, but it is easy to downgrade or eliminate that error
if it is a problem for users.
