+++ integrations/tensorflow/iree_tf_compiler/MHLO/BUILD
@@ -13,7 +13,6 @@ package(
 cc_library(
     name = "MHLO",
     srcs = [
-        "EmitDefaultIREEABI.cpp",
         "Passes.cpp",
     ],
     hdrs = [
