+++ iree/compiler/Bindings/Native/Transforms/test/emit_default_abi.mlir
@@ -1,4 +1,4 @@
-// RUN: iree-tf-opt %s -iree-mhlo-emit-default-iree-abi -split-input-file -verify-diagnostics | FileCheck %s
+// RUN: iree-opt %s -iree-abi-emit-default-abi -split-input-file -verify-diagnostics | FileCheck %s
 
 // CHECK-LABEL: func @valid
 // CHECK-SAME{LITERAL}: iree.abi = "{\22a\22:[[\22ndarray\22,\22f32\22,2,2,3],[\22ndarray\22,\22f32\22,1,3]],\22r\22:[[\22ndarray\22,\22f32\22,1,3],[\22ndarray\22,\22f32\22,2,2,3]],\22v\22:1}"
